+++ Gemfile
@@ -64,5 +64,4 @@ gem 'pg'
 gem 'figaro'
 gem 'rails_12factor', group: :production
 
-gem 'differ'
 gem 'diffy'
